Add anti‐aliasing, for no good reason but smoothness
Ha, found this patch on my HDD from 2011-06-20, though this little
project is defunct, there’s no reason not to commit this and rm the
patchfile. I didn’t understand much about webgl when I started this back
then, so this kinda embarasses me now, but what the hey. No doubt I will
replace it in the future.